+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@ Version 1.1.0 (Tamarack): Not yet released
 - Fix listen server create panel not properly loading `Level Name` field for Alpine levels
 - Fix geomod autotexture ppm not being correctly calculated for all crater textures
 - Fix "Time Left" text label not being localized
+- Fix `Player Entity Type`, `Player Undercover Suit Entity Type`, `Player Undercover Scientist Entity Type` not always applying
 
 ### New features and enhancements
 [@GooberRF](https://github.com/GooberRF)
@@ -116,7 +117,7 @@ Version 1.1.0 (Tamarack): Not yet released
 - Add `sp_camerashake` command to turn off all instances of camera shake
 - Add `mp_playernames` command to toggle drawing of names of targeted players in multiplayer
 - Add `mp_verbosetimer` command to toggle drawing of the "Time Left" label on the multiplayer timer
-- Add `Multi Timer X Offset`, `Multi Timer Y Offset`, and `Multi Timer Color` to `af_client.tbl`_
+- Add `Multi Timer X Offset`, `Multi Timer Y Offset`, and `Multi Timer Color` to `af_client.tbl`
 
 [@natarii](https://github.com/natarii)
 - Add new damage notification hit sound file